Added support for first-word, non-exact matches
When a query consists of only one word, the poster most likely wanted to link a card whose name starts with that word (instead of the word being in the middle of the name).

Example:
https://reddit.com/r/Shadowverse/comments/crutbh/i_fundamentally_dont_understand_this_game/ex9uqpd/

This commit refactors non-exact matches to first search for cards starting with that word, then resort to the usual methods when such card is not found.